refactor: request DTO 네이밍 개선
