+++ data/src/main/java/com/housweet/data/api/ReportApi.kt
@@ -2,7 +2,7 @@ package com.housweet.data.api
 
 import com.housweet.data.BuildConfig
 import com.housweet.data.network.KtorService
-import com.housweet.data.network.dto.ReportRequest
+import com.housweet.data.request.ReportRequest
 import com.housweet.data.network.dto.ReportResponse
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
