+++ data/src/main/java/com/housweet/data/network/HouseRegisterRemoteDataSource.kt
@@ -1,7 +1,7 @@
 package com.housweet.data.network
 
-import com.housweet.data.model.request.RegisterHouseRequest
-import com.housweet.data.network.dto.UpdateHouseRequest
+import com.housweet.data.request.RegisterHouseRequest
+import com.housweet.data.request.UpdateHouseRequest
 
 interface HouseRegisterRemoteDataSource {
     suspend fun registerHouse(body: RegisterHouseRequest)
