+++ data/src/main/java/com/housweet/data/network/ReportRemoteDataSource.kt
@@ -1,7 +1,7 @@
 package com.housweet.data.network
 
 import com.housweet.data.api.ReportApi
-import com.housweet.data.network.dto.ReportRequest
+import com.housweet.data.request.ReportRequest
 import com.housweet.data.network.dto.ReportResponse
 import javax.inject.Inject
 import javax.inject.Singleton
