+++ data/src/main/java/com/housweet/data/repository/HouseRegisterRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package com.housweet.data.repository
 
-import com.housweet.data.model.request.RegisterHouseRequest
+import com.housweet.data.request.RegisterHouseRequest
 import com.housweet.data.network.HouseRegisterRemoteDataSource
-import com.housweet.data.network.dto.UpdateHouseRequest
+import com.housweet.data.request.UpdateHouseRequest
 import com.housweet.data.network.dto.toDomain
 import com.housweet.domain.model.HouseRegisterModel
 import com.housweet.domain.model.RoomPostingDetail
