+++ data/src/main/java/com/housweet/data/repository/ReportRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.housweet.data.repository
 
 import com.housweet.data.network.ReportRemoteDataSource
-import com.housweet.data.network.dto.ReportRequest
+import com.housweet.data.request.ReportRequest
 import com.housweet.domain.model.Report
 import com.housweet.domain.repository.ReportRepository
 import javax.inject.Inject
