+++ data/src/main/java/com/housweet/data/request/BlockUserRequest.kt
@@ -1,10 +1,10 @@
-package com.housweet.data.dto
+package com.housweet.data.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class BlockUserRequestDto(
+data class BlockUserRequest(
     @SerialName("blocked_user_id")
     val blockedUserId: Int
 )
