+++ data/src/main/java/com/housweet/data/request/CreateRoomRequest.kt
@@ -1,4 +1,4 @@
-package com.housweet.data.network.dto
+package com.housweet.data.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
