+++ data/src/main/java/com/housweet/data/request/ProfileUpdateRequest.kt
@@ -1,13 +1,18 @@
-package com.housweet.data.dto
+package com.housweet.data.request
 
 import com.housweet.domain.model.profile.ProfileUpdateModel
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
 
-data class ProfileUpdateDto(
+@Serializable
+data class ProfileUpdateRequest(
     val gender: String,
     val introduce: String,
     val mbti: String,
     val nickname: String,
+    @SerialName("tag")
     val tags: List<String>,
+    @SerialName("year_of_birth")
     val yearOfBirth: String
 ) {
     fun mapToProfileUpdateModel(): ProfileUpdateModel {
