+++ data/src/main/java/com/housweet/data/request/RegisterHouseRequest.kt
@@ -1,21 +1,16 @@
-package com.housweet.data.network.dto
+package com.housweet.data.request
 
-import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class UpdateHouseRequest(
+data class RegisterHouseRequest(
     val room: Int,
-//    val si: Int,
-//    val gu: Int,
-//    val dong: Long,
     val si: String,
     val gu: String,
     val dong: String,
     val title: String,
     val content: String,
     val image_uri: String,
-//    val images: List<String>,
     val traffic_tags: List<String>,
     val size_of_house_tags: List<String>,
     val infra_tags: List<String>,
@@ -25,5 +20,5 @@ data class UpdateHouseRequest(
     val rent: Int,
     val deposit: Int,
     val management_fee: Int,
-    @SerialName("available_from") val available_from: String
+    val available_from: String
 )
