+++ data/src/main/java/com/housweet/data/request/UpdateMoodRequest.kt
@@ -1,10 +1,10 @@
-package com.housweet.data.dto
+package com.housweet.data.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class UpdateMoodRequestDto(
+data class UpdateMoodRequest(
     @SerialName("feeling")
     val feeling: String
 )
