+++ data/src/main/java/com/housweet/data/request/VisibilityRequest.kt
@@ -1,9 +1,9 @@
-package com.housweet.data.network.dto
+package com.housweet.data.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class VisibilityRequestDto(
+data class VisibilityRequest(
     @SerialName("is_visible") val isVisible: Boolean
 )
